add filename val,  return types, resize tests

+++ src/tests/utilities/sharpUtilSpec.ts
@@ -3,6 +3,7 @@ import app from '../..';
 import fs from 'fs';
 import path from 'path';
 import resize from '../../utilities/sharpUtil';
+import { doesNotMatch } from 'assert';
 
 const request = supertest(app);
 
@@ -11,22 +12,32 @@ const validParams = {
     width: 500,
     height: 250,
 };
+const invalidParams = {
+    filename: 'x',
+    width: -200,
+    height: 0,
+    extraParam: 'XYZ',
+};
 const testImagePath = path.join(
     __dirname,
     '../../../images/output/',
     `${validParams.filename}-${validParams.width}x${validParams.height}.jpg`
 );
 
+beforeEach(function () {
+    try {
+        if (fs.existsSync(testImagePath)) {
+            fs.unlink(testImagePath, (err) => {
+                if (err) return console.warn(`Image not found`);
+                console.warn('Test image deleted');
+            });
+        }
+    } catch {
+        throw new Error();
+    }
+});
+
 describe('sharpUtil tests:', () => {
-    afterAll(function () {
-        fs.unlink(testImagePath, (err) => {
-            if (err) console.log(err);
-            else {
-                console.warn('\ntesting image deleted');
-            }
-        });
-    });
-    
     it('Expect generation of image via end point.', async () => {
         await request.get(
             `/api/images?filename=${validParams.filename}&width=${validParams.width}&height=${validParams.height}`
@@ -34,14 +45,20 @@ describe('sharpUtil tests:', () => {
         expect(fs.existsSync(testImagePath)).toBeTrue();
     });
     it('Expect generation of image via sharp', async () => {
-
-        expect(async () => {
-            await resize(
-                validParams.filename,
-                validParams.width,
-                validParams.height
-            );
-        }).not.toThrow();
+        await resize(
+            validParams.filename,
+            validParams.width,
+            validParams.height
+        );
         expect(fs.existsSync(testImagePath)).toBeTrue();
+    });
+    it('Expect error on image generation via sharp', async () => {
+        await expectAsync(
+            resize(
+                invalidParams.filename,
+                invalidParams.width,
+                invalidParams.height
+            )
+        ).toBeRejectedWithError();
     });
 });
